JSON parsing of basic Location object
Note that this is not yet tested, as I do not yet have a server at my
disposal to produce the JSON that I want.  However, there is basic logic
in this commit to allow for the parsing of a JSON object representing a
Location with a name and an ID field to be parsed into a POJO.

Once I have the Rails server accessible I will be able to legitimately
continue developing on this branch.