SessionService initial implementation
Also attempted to get a Snackbar to show persistently in the LoginActivity
whilst not connected to the internet.  It is failing (I think) because
when the broadcast is received indicating a
ConnectivityManager.CONNECTIVITY_ACTION the network connection is in some
in-between state.  We can't use this to show or dismiss the Snackbar.

I am commiting this now so that it is in the history, but will delete the
broadcasting logic and instead opt to just check the connection when the
button to log in is pressed before performing the login.  If not connected
to the internet, I'll display a Snackbar to this effect.