Checking with SessionService if we are logged in.
SessionService checks with the session status that is stored in
SharedPreferences.  Ideally SessionService will be checking somehow that
we are authenticated with the server.  This might mean that the
authenitcation status is stored in a static boolean (dangerous) or might
mean that we check with the server.

In any case, we are going to need to work on the code server-side before
we can make any worthwhile progress with login code on the client-side.