SplashScreenActivity dynamically sends user to Login or Main
This required that SplashScreenActivity extend SessionAwareActivity and in
doing so also "override" the member variable that defines the service
connection so that we could perform a check for whether a user is logged
in in the onServiceConnected() callback.

This all happens very rapidly, so we barely even see the splash screen...
perhaps we should have some default time that the screen displays anyway,
regardless of whether we know the user is logged in or not.