Timer for splash re-implemented, with login checking
Note that the behaviour here all depends on the behaviour of
SessionService and its loggedIn() boolean method.  If this method takes a
long time to return, we may have an issue and/or even if the user *is*
logged in they may be sent to LoginActivity.

This will need to be remedied at a lated time, using a broadcast.